Use round line joins (matters for sharp corners)
