+++ docs/index.html
@@ -21,6 +21,7 @@
 	<script type="module" src="game.js"></script>
 	<script type="module" src="kart.js"></script>
 	<script type="module" src="math.js"></script>
+	<script type="module" src="track.js"></script>
 </body>
 
 </html>
