+++ game.ts
@@ -1,6 +1,7 @@
 import { Bumper } from "./bumper.js";
 import { Kart } from "./kart.js";
 import { Angle, clamp, mod, Vec2 } from "./math.js";
+import { Corner, Track } from "./track.js"
 
 const UPDATES_PER_SEC = 60.0;
 const MS_PER_UPDATE = 1000.0 / UPDATES_PER_SEC;
@@ -16,7 +17,6 @@ const WALL_BOUNCE_LOSS = 0.3;
 const MAX_STEERING_ANGLE = Math.PI / 6.0;
 
 const TRACK_RADIUS = 50.0;
-const TRACK_BORDER = 2.0;
 
 // MouseAxes control scheme constants
 
@@ -40,107 +40,6 @@ enum ControlScheme {
 	MouseAxes,
 	MouseFollow,
 	GamepadFollow,
-}
-
-class Corner {
-	vertex: Vec2;
-	// Values outside [0, 1] may result in loops.
-	smoothness: number;
-
-	constructor(vertex: Vec2, smoothness: number) {
-		this.vertex = vertex;
-		this.smoothness = smoothness;
-	}
-}
-
-class CubicBezier {
-	start: Vec2;
-	end: Vec2;
-	cp1: Vec2;
-	cp2: Vec2;
-
-	constructor(start: Vec2, end: Vec2, cp1: Vec2, cp2: Vec2) {
-		this.start = start;
-		this.end = end;
-		this.cp1 = cp1;
-		this.cp2 = cp2;
-	}
-}
-
-class Track {
-	private name: string;
-	private radius: number;
-	private spline: CubicBezier[];
-
-	constructor(name: string, radius: number, corners: Corner[]) {
-		this.name = name;
-		this.radius = radius;
-		// Build a list of unit offset vectors from each vertex to its control
-		// point in the forward direction.
-		let forwardCPOffsets: Vec2[] = [];
-		for (let i = 0; i < corners.length; ++i) {
-			const v = corners[i].vertex;
-			const vPrev = corners[(i - 1 + corners.length) % corners.length].vertex;
-			const vNext = corners[(i + 1) % corners.length].vertex;
-			const fromPrev = v.minus(vPrev);
-			const toNext = vNext.minus(v);
-			// Get a vector bisecting the angle of the corner.
-			const bisector = fromPrev.times(toNext.length()).plus(toNext.times(-fromPrev.length()));
-			// Use the z-coordinate of the cross product to determine if it's a
-			// left or right turn and therefore which way to rotate the offset.
-			const crossZ = fromPrev.x * toNext.y - fromPrev.y * toNext.x;
-			const rotated = crossZ > 0 ? bisector.rotatedQuarter() : bisector.rotatedThreeQuarters();
-			forwardCPOffsets.push(rotated.normalized());
-		}
-
-		this.spline = [];
-		for (let i = 0; i < corners.length; ++i) {
-			const next = (i + 1) % corners.length;
-			const start = corners[i].vertex;
-			const end = corners[next].vertex;
-			const l = end.minus(start).length() / 3;
-			const cp1 = start.plus(forwardCPOffsets[i].times(corners[i].smoothness * l));
-			const cp2 = end.minus(forwardCPOffsets[next].times(corners[next].smoothness * l));
-			this.spline.push(new CubicBezier(start, end, cp1, cp2));
-		}
-	}
-
-	draw(debug: boolean) {
-		this.drawSplines(this.radius, "black");
-		this.drawSplines(this.radius - TRACK_BORDER, "rgb(60, 60, 60)");
-
-		if (debug) {
-			ctx.font = "20pt serif";
-			ctx.fillStyle = "white";
-			ctx.fillText(this.name, 10, 30);
-			// Draw Bezier curve "frames".
-			ctx.lineWidth = 1;
-			let even = true;
-			for (let curve of this.spline) {
-				ctx.strokeStyle = even ? "red" : "white";
-				even = !even;
-				ctx.beginPath();
-				ctx.moveTo(curve.start.x, curve.start.y);
-				ctx.lineTo(curve.cp1.x, curve.cp1.y);
-				ctx.lineTo(curve.cp2.x, curve.cp2.y);
-				ctx.lineTo(curve.end.x, curve.end.y);
-				ctx.stroke();
-			}
-		}
-	}
-
-	private drawSplines(radius: number, style: string) {
-		ctx.beginPath();
-		ctx.strokeStyle = style;
-		const start = this.spline[0].start;
-		ctx.moveTo(start.x, start.y);
-		for (let curve of this.spline) {
-			ctx.bezierCurveTo(curve.cp1.x, curve.cp1.y, curve.cp2.x, curve.cp2.y, curve.end.x, curve.end.y);
-			ctx.lineWidth = 2 * radius;
-		}
-		ctx.closePath();
-		ctx.stroke();
-	}
 }
 
 const tracks: Track[] = [
@@ -380,7 +279,7 @@ class MainScene {
 		ctx.rect(0, 0, canvas.width, canvas.height);
 		ctx.fill();
 
-		tracks[this.trackIdx].draw(this.debug);
+		tracks[this.trackIdx].draw(ctx, this.debug);
 
 		// Draw walls.
 		for (const wall of this.walls) {
